refactor: 공통 응답(PageResponse) 구조 변경
