+++ src/main/java/com/sparta/todayeats/area/presentation/controller/AreaController.java
@@ -5,7 +5,7 @@ import com.sparta.todayeats.area.presentation.dto.AreaCreateRequest;
 import com.sparta.todayeats.area.presentation.dto.AreaCreateResponse;
 import com.sparta.todayeats.area.presentation.dto.AreaResponse;
 import com.sparta.todayeats.area.presentation.dto.AreaUpdateRequest;
-import com.sparta.todayeats.category.presentation.dto.PageResponse;
+import com.sparta.todayeats.global.response.PageResponse;
 import com.sparta.todayeats.global.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
